"Which version takes which options" into a module
Here and there I saw ad-hoc functions for deciding which
contracts_version expects which options.

This commit gathers these functions into a new module.

This makes CodeClimate happier https://github.com/raiden-network/raiden-contracts/issues/704
